requirements.txt
beautifulsoup4==4.12.3
greenlet==3.0.3
logging==0.4.9.6
numpy==1.26.4
pandas==2.2.2
playwright==1.44.0
psycopg2==2.9.9
pyee==11.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.1
tzdata==2024.1
